% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorption_order.R
\name{AbsOrderPH}
\alias{AbsOrderPH}
\title{Quantities used in calculation of absorption order probabilities.}
\usage{
AbsOrderPH(listS, listAlpha)
}
\arguments{
\item{listS}{list of sub-intensity matrices of the Phase-types.}

\item{listAlpha}{list of initial distribution vectors of the Phase-types.}
}
\value{
The vector \code{pi, T, D} of theorem 4.23.
}
\description{
Quantities used in calculation of absorption order probabilities.
}
