% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorption_order.R
\name{dimAbsOrderPH}
\alias{dimAbsOrderPH}
\title{Dimension of Absorption order matrix}
\usage{
dimAbsOrderPH(dimVec)
}
\arguments{
\item{dimVec}{vector of dimensions of the phase-types.}
}
\value{
The corresponding dimension of the D matrix for calculating the absorption order probabilities.
}
\description{
Dimension of Absorption order matrix
}
