% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimOrderStat.R
\name{dimOrderStat}
\alias{dimOrderStat}
\title{The dimension of the k'th order stat of n PH}
\usage{
dimOrderStat(k, n, pvec)
}
\arguments{
\item{k}{the order.}

\item{n}{total number of phase-types from which to calculate dimension of kth order statistic.}

\item{pvec}{vector of dimensions of n phase-types. If all equal can be single number.}
}
\value{
Dimension of the phase-type representation of the kth order statistic of n.
}
\description{
The dimension of the k'th order stat of n PH
}
