% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscretePHGenerator.R
\name{DPHSubSeqGen}
\alias{DPHSubSeqGen}
\title{Discrete Phase-type representation for first occurrence of k-sub sequence.}
\usage{
DPHSubSeqGen(pvec = c(0.8, 0.2), k_sub_seq = c(1, 1, 2, 2))
}
\arguments{
\item{pvec}{probability vector of unique outcomes}

\item{k_sub_seq}{sequence to be implemented as DPH. Must start with 1 and increment by one for each new unique outcome.}
}
\value{
Initial vector and sub-transition matrix of the discrete phase-type distribution of the first occurrence.
}
\description{
\code{DPHSubSeqGen} must have the sub-sequence supplied in the following way: Start with a 1, then for every new unique value increment by one
otherwise repeat the value.
}
\examples{
##For the first occurrence of the sequence (10,9,8,8,9,10) of iid Pois(7) random variables.
DPHSubSeqGen(pvec = dpois(c(10,9,8), lambda = 7), k_sub_seq = c(1,2,3,3,2,1))
}
