% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang_generator.R
\name{erlang_ph}
\alias{erlang_ph}
\title{Erlang Feedback example
General Erlang (feedback) generator}
\usage{
erlang_ph(k = 10, rate = 1:10, feedback = NULL)
}
\arguments{
\item{k}{number of phases in the underlying process.}

\item{rate}{rate parameter. Can also be a k vector.}

\item{feedback}{Feedback probability parameter, default value is zero.}
}
\description{
Erlang Feedback example
General Erlang (feedback) generator
}
