% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{ImpSampler}
\alias{ImpSampler}
\title{Importance Sampler of process underlying a Phase-type distribution conditioned on absorption time.}
\usage{
ImpSampler(pi = listAlpha[["pi4"]], SubIntMat = listS[["S4"]],
  N = 100, TimePoint = 0.8, OFun = OtimeSpent, ...)
}
\arguments{
\item{pi}{Initial distribution of underlying process.}

\item{SubIntMat}{Sub-intensity matrix of underlying process.}

\item{N}{Number of total iterations.}

\item{TimePoint}{Time of absorption to be conditioned on.}

\item{OFun}{Functional applied to underlying process.}

\item{...}{Additional argument passed on to \code{OFun}.}
}
\description{
Importance Sampler of process underlying a Phase-type distribution conditioned on absorption time.
}
\examples{
ImpSampler(State = 1)
}
