% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{MCMC_PH}
\alias{MCMC_PH}
\title{MCMC}
\usage{
MCMC_PH(pi = ErInt, SubIntMat = ErMat, N = 100, BurnIn = NULL,
  ForcePosExitRateStart = F, TimePoint = 0.8, OFun = OtimeSpent, ...)
}
\arguments{
\item{pi}{The initial distribution of the underlying Markov jump process.}

\item{SubIntMat}{The sub-intensity matrix of the underlying Markov jump process.}

\item{N}{Number of steps in the MCMC algorithm.}

\item{BurnIn}{Burn In.}

\item{ForcePosExitRateStart}{Logical, should the first chain be forced to have positive exit rate at end state? Default is FALSE.}

\item{TimePoint}{Time of absorption conditioned on.}

\item{OFun}{The functional of the underlying process.}

\item{...}{additional arguments to be passed on to OFun.}
}
\value{
Estimated mean of function \code{Ofun} applied to the underlying process.
}
\description{
Markov Chain Monte Carlo Estimator of functional of underlying Markov jump process underlying a Phase-type distribution
conditioned on the absorption time being equal to x = TimePoint.
}
\examples{
MCMC_PH(pi = listAlpha[[4]], SubInt = listS[[4]], N = 1e4, BurnIn = 1e2, TimePoint = 0.8, OFun = OtimeSpent, State = 5)
}
