% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_functions.R
\name{OcondCDF}
\alias{OcondCDF}
\title{Functional of Markov jump process}
\usage{
OcondCDF(obj, State, y)
}
\arguments{
\item{obj}{Markov jump process object with state and jump sequence and intensity matrix.}

\item{State}{State in the state space of the Markov jump process.}

\item{y}{}
}
\value{
1 if the time spent in state \code{State} is less than or equal to \code{y} 0 else.
}
\description{
Use for estimating conditional cdf given absorption time.
}
\examples{
(a <- simPH(SubIntMat = listS[[6]], phi0 = listAlpha[[6]]))
OcondCDF(a, State = 3, y = 0.2)
}
