% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_functions.R
\name{OtimeSpent}
\alias{OtimeSpent}
\alias{Ojumpsij}
\alias{OtotalNrJumps}
\title{O-functions
Time spent in state \code{i}}
\usage{
OtimeSpent(obj, State)

Ojumpsij(obj, i = 1, j = 2)

OtotalNrJumps(obj)
}
\arguments{
\item{obj}{A Markov jump process object with state and jump sequence and intensity matrix.}

\item{State}{}

\item{i}{A state in the state space of the Markov jump process.}

\item{j}{A state in the state space of the Markov jump process.}

\item{obj}{}
}
\value{
The total number of jump from state \code{i} to \code{j} of the Markov jump process given by \code{obj}.
}
\description{
O-functions
Time spent in state \code{i}
}
\section{Functions}{
\itemize{
\item \code{Ojumpsij}: total number of jump from \eqn{i -> j}

\item \code{OtotalNrJumps}: total number of jump
}}

\note{
Note last state in \code{RejMJP} is not an actual jump

Note last state in \code{RejMJP} is not an actual jump and first state is not a jump either
}
\examples{
(a <- RejMJP(TimePoint = 1))
OtimeSpent(a, State = 2)

(obj <- epsRejMJP(TimePoint = 0.6))
OtimeSpent(obj, State = 2)
(obj <- RejMJP(TimePoint = 6))
Ojumpsij(obj, i = 2, j = 1)
(obj <- RejMJP(TimePoint = 1))
OtotalNrJumps(obj = obj)
}
