% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_jump_rejection_sampling.R
\name{RejMJP}
\alias{RejMJP}
\title{Rejection sample a Markov jump process still active at time = Timepoint}
\usage{
RejMJP(IntDist = listAlpha[[6]], SubIntMat = listS[[6]],
  TimePoint = 0.8)
}
\arguments{
\item{IntDist}{Initial distribution of the Markov jump process.}

\item{SubIntMat}{Sub-intensity matrix of the Markov jump process.}

\item{TimePoint}{Timepoint at which the Markov jump process must surpass.}
}
\value{
Markov jump process active at time x=Timepoint. Including Sequence of visited states and holding times.
}
\description{
Rejection sample a Markov jump process still active at time = Timepoint
}
