% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{simMC}
\alias{simMC}
\title{Function for simulation Markov Chains and Markov Jump Process}
\usage{
simMC(tr, nJump = 10, phi0 = NULL, RunUntilAbs = FALSE)
}
\arguments{
\item{RunUntilAbs}{}
}
\description{
Function for simulation Markov Chains and Markov Jump Process
}
