% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{simMPH}
\alias{simMPH}
\title{MPH* simulation}
\usage{
simMPH(SubIntMat, IntVec = NULL, RewardMat)
}
\arguments{
\item{RewardMat}{}
}
\description{
MPH* simulation
}
\note{
Note IntVec and IntMat should be p dimensional, for the underlying chain on $E = 1,2,...,p,p+1$
}
