% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{simPH}
\alias{simPH}
\title{SimPH , wrapper for simMC, adding the p+1 state and letting the chain run until absorption}
\usage{
simPH(SubIntMat = listS[["S3"]], phi0 = NULL)
}
\arguments{
\item{phi0}{}
}
\description{
SimPH , wrapper for simMC, adding the p+1 state and letting the chain run until absorption
}
