% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimOrderStat.R
\name{sumKronecker}
\alias{sumKronecker}
\title{Kronecker SUM}
\usage{
sumKronecker(A, B)
}
\arguments{
\item{A}{Square matrix.}

\item{B}{Square matrix.}
}
\value{
The Kronecker Sum between matrices A and B.
}
\description{
Kronecker SUM
}
\examples{
sumKronecker(listS[[1]], listS[[2]])
}
