% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Urn3.R
\name{Urn3}
\alias{Urn3}
\title{Discrete phase-type representation of Urn experiment as given in Example 3.18}
\usage{
Urn3(n, d, k)
}
\arguments{
\item{n}{total number of balls in urn}

\item{d}{number of black balls in urn}

\item{k}{k'th black ball}
}
\value{
Phase type representation (\code{pi_kn} , \code{T_kn}) as a list.
}
\description{
Discrete phase-type representation of Urn experiment as given in Example 3.18
}
\examples{

Urn3(100, 25, 10)
}
