% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasetypeUtilsUcph-package.R
\docType{package}
\name{phasetypeUtilsUcph-package}
\alias{phasetypeUtilsUcph}
\alias{phasetypeUtilsUcph-package}
\title{phasetypeUtilsUcph: Phase-type Distribution Helpers}
\description{
Utility functions for phase-type distributions.
}
\author{
\strong{Maintainer}: Niels Krarup \email{krarup.niels@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
