% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Urn3.R
\name{Urn3}
\alias{Urn3}
\title{Urn Example 3}
\usage{
Urn3(n, d, k)
}
\arguments{
\item{n}{total number of balls in urn}

\item{d}{number of black balls in urn}

\item{k}{k'th black ball}
}
\description{
Urn Example 3
}
\examples{

Urn3(100, 25, 10)
}
